use signed type for getfilesize
both windows and linux use signed types for this stat so it's better to use signed here

+++ myIO.c
@@ -22,15 +22,15 @@ void mymkdir(const char *const path) {
     }
 }
 
-size_t getfilesize(const char *const filePath) {
+intmax_t getfilesize(const char *const filePath) {
 #ifdef _WIN32
     _stat64 sb;
     const int returnValue = _wstat64(filePath, &sb);
-    const size_t size = (size_t) sb.st_size;
+    const intmax_t size = (intmax_t) sb.st_size;
 #else
     struct stat sb;
     const int returnValue = stat(filePath, &sb);
-    const size_t size = (size_t) sb.st_size;
+    const intmax_t size = (intmax_t) sb.st_size;
 #endif
     if (returnValue != 0) {
         perror("ERROR: Failed to get file size");
